requirements.txt
curl-cffi>=0.6.0
python-dotenv>=1.0.0
